check multipass at more places

+++ bin/fuzz.py
@@ -70,7 +70,7 @@ class GeneratorTask:
         localbar.done()
 
         # Generate .ans.
-        if not self.fuzz.problem.interactive:
+        if not self.fuzz.problem.interactive and not self.fuzz.problem.multipass:
             if self.solution and not testcase.ans_path.is_file():
                 if testcase.ans_path.is_file():
                     testcase.ans_path.unlink()
